Experimentation with spawning a worker reusing the current WASM module
This might work, but needs a modified version of wasm-bindgen. The main problem is providing the imports to the module when instantiating it in the worker. Probably requires generating a secondary JS "init" file for the worker. The "spawn worker" function would have to be aware of the generated file name (which probably means it would have to be a custom-generated function as well, I don't think that's feasible with a normal `#[wasm_bindgen]` function; actually, as a write this, including this file inline as a string constant and then initiating the worker with a ObjectUrl from a Blob that contains this string may be able to accomplish this).

Another open question is if work needs to be done to setup the "shadow stack" for the worker, or whether instantiating the module again will do that automatically.